LayoutTests/fast/events/onunload-window-property-expected.txt

ALERT: unload
You should have seen an unload alert appear.
